requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
chromadb==0.5.0
sentence-transformers==3.0.1
PyMuPDF==1.24.4
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.3
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
bcrypt==4.1.3
groq==1.1.1
numpy<2.0.0
gunicorn==21.2.0
